Code clean ups, update to 1.14.4

+++ src/main/java/net/dzikoysk/wildskript/util/WildSkriptTimer.java
@@ -8,7 +8,7 @@ public class WildSkriptTimer implements Runnable {
     private final LinkedList<Double> history = new LinkedList<>();
 
     public WildSkriptTimer() {
-        this.history.add(Double.valueOf(20.0D));
+        this.history.add(20.0D);
     }
 
     public void run() {
@@ -17,13 +17,13 @@ public class WildSkriptTimer implements Runnable {
         if (timeSpent == 0L) timeSpent = 1L;
         if (this.history.size() > 10) this.history.remove();
         double tps = 50000000.0D / timeSpent;
-        if (tps <= 21.0D) this.history.add(Double.valueOf(tps));
+        if (tps <= 21.0D) this.history.add(tps);
         this.lastPoll = startTime;
     }
 
     public double getTPS() {
         double avg = 0.0D;
-        for (Double f : this.history) if (f != null) avg += f.doubleValue();
+        for (Double f : this.history) if (f != null) avg += f;
         return avg / this.history.size();
     }
 }
